refactor: `MemoItem` 컴포넌트 분리
